requirements.txt
Flask==2.3.2
tensorflow-cpu==2.12.0
scikit-learn==1.3.0
numpy==1.23.5 # Ensures this specific compatible version
pandas==2.0.3
matplotlib==3.7.2
joblib==1.3.2
gunicorn==21.2.0
flask-cors==4.0.0
wtforms==3.0.1
requests==2.31.0
boto3==1.28.57
